Add count to downlink 'pong' message
This commit adds the count from the received uplink payload, like
'ping:14', to the downlink response message which would become
'pong:14'.

The motivation for this is that it is easier to identify the which pong
is received in the device console output.

For example, instead of just seeing:

INFO  Received 4 bytes from uplink:
pong

The count will now also be included:

INFO  Received 7 bytes from uplink:
pong:14